Refactor route setup
